8bits

+++ proj3/crn-header.hh
@@ -6,8 +6,8 @@ CLICK_DECLS
 struct CrnPacket{
 	
 	uint8_t type;//0 for request, 1 for content and 2 for update 
-	uint32_t hopcount;
-	uint32_t content_id;
+	uint8_t hopcount;
+	uint8_t content_id;
 	in_addr my_interface;
 	const unsigned char* data;
